[TWINFACES-313] refactor: rename files and folders
