+++ src/screens/factory-branche/index.ts
@@ -1,0 +1,1 @@
+export * from "./factory-branche";
