--- src/screens/factoryBranches/index.ts
@@ -1,1 +1,0 @@
-export * from "./factoryBranches";
